+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## BuildBattle 3 changelog
 
+### 3.0.2 Release (03.06.2018)
+* Fixed /bb create command not working
+* Added /bba settheme command
+* Fixed game boss bar not removing after plugin force disable
+
 ### 3.0.0 Release (14.05.2018 - 01.06.2018)
 * Add everything here